feat: implement CSS-based LogoCarousel with reverse animation support
- Create new LogoCarousel component with CSS Modules animations
- Replace GSAP-based implementation with pure CSS keyframes for better portability
- Add reverse prop to control scroll direction (left-to-right or right-to-left)
- Fix component folder name from LogoCaurosel to LogoCarousel (typo correction)
- Implement hover pause/resume with animationPlayState
- Add comprehensive JSDoc documentation with examples
- Use CSS Modules for scoped animation styles (logoLoop and logoLoopReverse)
- Adjust default speed from 50 to 15 for smoother animation
- Increase logo height from h-12 to h-20 on mobile for better visibility
- Remove GSAP dependency from carousel (simpler, more portable solution)
