+++ src/components/LogoCarousel/styles.module.css
@@ -1,0 +1,27 @@
+/* Infinite horizontal loop animation for logo carousel */
+.carousel_animate_loop {
+  animation: logoLoop linear infinite;
+}
+
+/* Reverse direction animation (right to left) */
+.carousel_animate_loop_reverse {
+  animation: logoLoopReverse linear infinite;
+}
+
+@keyframes logoLoop {
+  from {
+    transform: translate(-100%);
+  }
+  to {
+    transform: translate(0);
+  }
+}
+
+@keyframes logoLoopReverse {
+  from {
+    transform: translate(0);
+  }
+  to {
+    transform: translate(-100%);
+  }
+}
